Fixing merge errors
